create package for mapper and dtos of deals, do refactoring of check methods in rentalService, move some exceptions in the right palce
