+++ src/main/java/com/develhope/spring/configuration/customAnnotations/EmailValidator.java
@@ -1,4 +1,4 @@
-package com.develhope.spring.configuration.validators;
+package com.develhope.spring.configuration.customAnnotations;
 
 
 import jakarta.validation.ConstraintValidator;
