+++ src/main/java/com/develhope/spring/deals/controllers/RentalController.java
@@ -1,8 +1,8 @@
 package com.develhope.spring.deals.controllers;
 
-import com.develhope.spring.deals.dtos.RentalCreatorDTO;
-import com.develhope.spring.deals.dtos.RentalReturnerDTO;
-import com.develhope.spring.deals.dtos.RentalUpdaterDTO;
+import com.develhope.spring.deals.dtos.rentalsDtos.RentalCreatorDTO;
+import com.develhope.spring.deals.dtos.rentalsDtos.RentalReturnerDTO;
+import com.develhope.spring.deals.dtos.rentalsDtos.RentalUpdaterDTO;
 import com.develhope.spring.deals.services.RentalService;
 import com.develhope.spring.users.models.User;
 import org.springframework.beans.factory.annotation.Autowired;
