--- src/main/java/com/develhope/spring/deals/dtos/OrderStatusDTO.java
@@ -1,23 +1,0 @@
-package com.develhope.spring.deals.dtos;
-
-import com.develhope.spring.deals.models.OrderStatus;
-
-public class OrderStatusDTO {
-
-    private OrderStatus orderStatus;
-
-    public OrderStatusDTO() {
-    }
-
-    public OrderStatusDTO(OrderStatus orderStatus) {
-        this.orderStatus = orderStatus;
-    }
-
-    public OrderStatus getOrderStatus() {
-        return orderStatus;
-    }
-
-    public void setOrderStatus(OrderStatus orderStatus) {
-        this.orderStatus = orderStatus;
-    }
-}
