+++ src/main/java/com/develhope/spring/users/models/User.java
@@ -2,7 +2,7 @@ package com.develhope.spring.users.models;
 
 import com.develhope.spring.deals.models.Order;
 import com.develhope.spring.deals.models.Rental;
-import com.develhope.spring.configuration.validators.CustomAnnotation.ValidEmail;
+import com.develhope.spring.configuration.customAnnotations.CustomAnnotation.ValidEmail;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
