+++ src/test/java/com/develhope/spring/deals/components/OrderMapperTest.java
@@ -1,7 +1,8 @@
 package com.develhope.spring.deals.components;
 
-import com.develhope.spring.deals.dtos.OrderCreatorDTO;
-import com.develhope.spring.deals.dtos.OrderResponseDTO;
+import com.develhope.spring.deals.components.mappers.OrderMapper;
+import com.develhope.spring.deals.dtos.ordersDtos.OrderCreatorDTO;
+import com.develhope.spring.deals.dtos.ordersDtos.OrderResponseDTO;
 import com.develhope.spring.deals.models.Order;
 import com.develhope.spring.deals.models.OrderStatus;
 import com.develhope.spring.users.components.UserMapper;
